開発環境用のFirestore collection追加

+++ operation/export.py
@@ -1,7 +1,7 @@
 import json
-import sys
 from pathlib import Path
 
+import click
 import firebase_admin
 from firebase_admin import credentials, firestore
 
@@ -12,18 +12,24 @@ app = firebase_admin.initialize_app(cred)
 client = firestore.client(app)
 
 
-def export_doc(path: Path, dt: str):
+@click.command()
+@click.argument("env")
+@click.argument("dt")
+def export_doc(env: str, dt: str):
+    if env == "prod":
+        if not click.confirm("are you sure you want to export into PROD?"):
+            click.echo("export is cancled")
+            return
+    path = env / Path("1103")
     ref = client.document(str(path))
     dir = "backup" / path
     with (dir / (dt + ".json")).open() as f:
         data = json.load(f)
     ref.delete()
-    ref.create({"data": data})
+    ref.create(data)
 
 
 # 実行例
-# $ poetry run python operation/export.py 202211190947
+# $ poetry run python operation/export.py dev 202211190947
 if __name__ == "__main__":
-    dt = sys.argv[1]
-    export_doc(Path("dev/decks"), dt)
-    export_doc(Path("dev/results"), dt)
+    export_doc()
